Create TicTacToe
Created a program capable of playing Tic-Tac-Toe against the user. Program uses OOP concepts and ASCII art to generate and display the 3x3 playing board. The program randomly decides who goes first (computer or user) and informs the user if an illegal move was made (cell already occupied). The program also announces
if one of the players wins or if a draw is achieved.