Create SymptomChecker
This program includes the header files iostream, ctime, & cstdlib for input-output operations, time functions, and random number generation. It declares the variables covid, cold, flu, and other to keep track of the number of patients with different illnesses. The class ‘info’ is defined with functions to print a symptom guide, get the number of patients, print stats, calculate the 
percentage of illnesses, & check for other illnesses. The class ‘specificInfo’ is defined & derived from ‘info’, which includes additional variables for symptoms & functions to assign symptoms and check illnesses based on symptoms.
The ‘main’ function includes:
-the random number generator with the current time
-prompting the user to input number of patients
-printing a symptom guide
-creating an array of ‘specificInfo’ based on number of patients entered
-generating random symptoms for each patient
-checking each patient's symptoms and puts them into different illnesses
-prints the stats of patients with each illness and their percentages
The symptoms are randomly assigned values (1, 2, or 3) for each patient, and based on combinations of symptoms, the program puts them into COVID-19, Cold, or Flu. All leftover patients are defined as having other illnesses. Then, it prints the percentages of patients with each illness.